Extract FeedLoader test helper into a shared protocol extension to remove duplication
